chore(swaps): cleanup
