+++ NArk.Swaps/Models/ArkSwapStatus.cs
@@ -1,0 +1,11 @@
+namespace NArk.Swaps.Models;
+
+public enum ArkSwapStatus
+{
+    Pending,
+    Settled,
+    Failed,
+    Recoverable,
+    Refunded,
+    Unknown
+}
