+++ NArk.Swaps/Models/ArkSwapType.cs
@@ -1,0 +1,9 @@
+namespace NArk.Swaps.Models;
+
+public enum ArkSwapType
+{
+    ReverseSubmarine,
+    Submarine,
+    ChainBtcToArk,
+    ChainArkToBtc
+}
