+++ NArk.Tests.End2End/SwapManagementServiceTests.cs
@@ -49,7 +49,7 @@ public class SwapManagementServiceTests
         var boltzProvider = new BoltzSwapProvider(boltzClient, new BoltzLimitsValidator(new CachedBoltzClient(new HttpClient(), new OptionsWrapper<BoltzClientOptions>(new BoltzClientOptions() { BoltzUrl = SharedSwapInfrastructure.BoltzEndpoint.ToString(), WebsocketUrl = SharedSwapInfrastructure.BoltzWsEndpoint.ToString() }))),
             testingPrerequisite.clientTransport, testingPrerequisite.vtxoStorage,
             testingPrerequisite.walletProvider, swapStorage, testingPrerequisite.contractService, testingPrerequisite.contracts,
-            testingPrerequisite.safetyService, spendingService, intentStorage, chainTimeProvider);
+            testingPrerequisite.safetyService, intentStorage, chainTimeProvider);
         await using var swapMgr = new SwapsManagementService(
             new ISwapProvider[] { boltzProvider },
             spendingService,
@@ -124,7 +124,7 @@ public class SwapManagementServiceTests
         var boltzProvider = new BoltzSwapProvider(boltzClient, new BoltzLimitsValidator(new CachedBoltzClient(new HttpClient(), new OptionsWrapper<BoltzClientOptions>(new BoltzClientOptions() { BoltzUrl = SharedSwapInfrastructure.BoltzEndpoint.ToString(), WebsocketUrl = SharedSwapInfrastructure.BoltzWsEndpoint.ToString() }))),
             testingPrerequisite.clientTransport, testingPrerequisite.vtxoStorage,
             testingPrerequisite.walletProvider, swapStorage, testingPrerequisite.contractService, testingPrerequisite.contracts,
-            testingPrerequisite.safetyService, spendingService, intentStorage, chainTimeProvider);
+            testingPrerequisite.safetyService, intentStorage, chainTimeProvider);
         await using var swapMgr = new SwapsManagementService(
             new ISwapProvider[] { boltzProvider },
             spendingService,
@@ -182,7 +182,7 @@ public class SwapManagementServiceTests
         var boltzProvider = new BoltzSwapProvider(boltzClient, new BoltzLimitsValidator(new CachedBoltzClient(new HttpClient(), new OptionsWrapper<BoltzClientOptions>(new BoltzClientOptions() { BoltzUrl = SharedSwapInfrastructure.BoltzEndpoint.ToString(), WebsocketUrl = SharedSwapInfrastructure.BoltzWsEndpoint.ToString() }))),
             testingPrerequisite.clientTransport, testingPrerequisite.vtxoStorage,
             testingPrerequisite.walletProvider, swapStorage, testingPrerequisite.contractService, testingPrerequisite.contracts,
-            testingPrerequisite.safetyService, spendingService, intentStorage, chainTimeProvider);
+            testingPrerequisite.safetyService, intentStorage, chainTimeProvider);
         await using var swapMgr = new SwapsManagementService(
             new ISwapProvider[] { boltzProvider },
             spendingService,
@@ -274,7 +274,7 @@ public class SwapManagementServiceTests
                 new OptionsWrapper<BoltzClientOptions>(new BoltzClientOptions()
                 { BoltzUrl = SharedSwapInfrastructure.BoltzEndpoint.ToString(), WebsocketUrl = SharedSwapInfrastructure.BoltzWsEndpoint.ToString() }))),
             prereq.clientTransport, prereq.vtxoStorage, prereq.walletProvider, swapStorage,
-            prereq.contractService, prereq.contracts, prereq.safetyService, spendingService,
+            prereq.contractService, prereq.contracts, prereq.safetyService,
             intentStorage, chainTimeProvider);
         await using var swapMgr = new SwapsManagementService(
             new ISwapProvider[] { boltzProvider },
@@ -441,7 +441,7 @@ public class SwapManagementServiceTests
                     new OptionsWrapper<BoltzClientOptions>(new BoltzClientOptions()
                     { BoltzUrl = SharedSwapInfrastructure.BoltzEndpoint.ToString(), WebsocketUrl = SharedSwapInfrastructure.BoltzWsEndpoint.ToString() }))),
                 prereq.clientTransport, prereq.vtxoStorage, prereq.walletProvider, swapStorage,
-                prereq.contractService, prereq.contracts, prereq.safetyService, spendingService,
+                prereq.contractService, prereq.contracts, prereq.safetyService,
                 intentStorage, chainTimeProvider);
             var swapMgr = new SwapsManagementService(
                 new ISwapProvider[] { boltzProvider },
@@ -520,7 +520,7 @@ public class SwapManagementServiceTests
                 { BoltzUrl = SharedSwapInfrastructure.BoltzEndpoint.ToString(), WebsocketUrl = SharedSwapInfrastructure.BoltzWsEndpoint.ToString() }))),
             testingPrerequisite.clientTransport, testingPrerequisite.vtxoStorage,
             testingPrerequisite.walletProvider, swapStorage, testingPrerequisite.contractService, testingPrerequisite.contracts,
-            testingPrerequisite.safetyService, spendingService, intentStorage, chainTimeProvider);
+            testingPrerequisite.safetyService, intentStorage, chainTimeProvider);
         await using var swapMgr = new SwapsManagementService(
             new ISwapProvider[] { boltzProvider },
             spendingService, testingPrerequisite.clientTransport, testingPrerequisite.vtxoStorage,
@@ -599,7 +599,7 @@ public class SwapManagementServiceTests
                     new OptionsWrapper<BoltzClientOptions>(new BoltzClientOptions()
                     { BoltzUrl = SharedSwapInfrastructure.BoltzEndpoint.ToString(), WebsocketUrl = SharedSwapInfrastructure.BoltzWsEndpoint.ToString() }))),
                 prereq.clientTransport, prereq.vtxoStorage, prereq.walletProvider, swapStorage,
-                prereq.contractService, prereq.contracts, prereq.safetyService, spendingService,
+                prereq.contractService, prereq.contracts, prereq.safetyService,
                 intentStorage, chainTimeProvider);
             var swapMgr = new SwapsManagementService(
                 new ISwapProvider[] { boltzProvider },
@@ -659,7 +659,7 @@ public class SwapManagementServiceTests
                     new OptionsWrapper<BoltzClientOptions>(new BoltzClientOptions()
                     { BoltzUrl = SharedSwapInfrastructure.BoltzEndpoint.ToString(), WebsocketUrl = SharedSwapInfrastructure.BoltzWsEndpoint.ToString() }))),
                 prereq.clientTransport, prereq.vtxoStorage, prereq.walletProvider, swapStorage,
-                prereq.contractService, prereq.contracts, prereq.safetyService, spendingService,
+                prereq.contractService, prereq.contracts, prereq.safetyService,
                 intentStorage, chainTimeProvider);
             var swapMgr = new SwapsManagementService(
                 new ISwapProvider[] { boltzProvider },
@@ -716,7 +716,7 @@ public class SwapManagementServiceTests
         var boltzProvider = new BoltzSwapProvider(boltzClient, new BoltzLimitsValidator(new CachedBoltzClient(new HttpClient(), new OptionsWrapper<BoltzClientOptions>(new BoltzClientOptions() { BoltzUrl = SharedSwapInfrastructure.BoltzEndpoint.ToString(), WebsocketUrl = SharedSwapInfrastructure.BoltzWsEndpoint.ToString() }))),
             testingPrerequisite.clientTransport, testingPrerequisite.vtxoStorage,
             testingPrerequisite.walletProvider, swapStorage, testingPrerequisite.contractService, testingPrerequisite.contracts,
-            testingPrerequisite.safetyService, spendingService, intentStorage, chainTimeProvider);
+            testingPrerequisite.safetyService, intentStorage, chainTimeProvider);
         await using var swapMgr = new SwapsManagementService(
             new ISwapProvider[] { boltzProvider },
             spendingService,
